+++ adapters/sequelize.js
@@ -24,6 +24,8 @@ module.exports = function (model, opts) {
       return 'Number';
     } else if (column.type instanceof DataTypes.JSONB) {
       return 'Json';
+    } else if (column.type instanceof DataTypes.TIME) {
+      return 'Time';
     } else if (column.type.type) {
       return [getTypeFor({ type: column.type.type })];
     }