+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Field Type - Support Time field type.
+- Models - Support resources/charts queries on models having a defaultScope configuration.
 
 ## RELEASE 0.3.7 - 2016-11-06
 ### Changed