+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 ### Fixed
 - Records Index - Fix the duplicate records displayed on the front if the ids are hidden in the list.
 
+### Fixed
+- Records Index - Fix the duplicate records displayed on the front if the ids are hidden in the list.
+
 ## RELEASE 0.3.7 - 2016-11-06
 ### Changed
 - BelongsTo - Better support belongsTo relationships (as, foreign keys, etc.).